TEI-2259 - Register browser log collection prior to first GET.
The addCleanups that register the function to collect the browser
logs in the selenium tests were erroneously placed after the first
GET. This means that if there is an error on that first GET, as in
the case of TEI-2013, we don't get the browser logs to debug the
issue. These cleanup functions need registered before we start
trying to use the driver in any way.

Change-Id: Icf93fd2552d1c64068a05cc89f694c503a7f00c9
Reviewed-on: http://review.whamcloud.com/10886
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>