HYD-3335: Use our pip cache script in 'make docs'.
Most of our build and test process has been made to use a cache
so that it won't fail if an outside dependency disappears or
has an outage. However, this was never extended to the docs make
target, which still does a full pip install every build, and
will fail the build if a requirement disappears. This extends our
pip cache script we use in the test jobs to the docs target.

Change-Id: I81e01044faa47e92b07e4ae1b439a9de4209a7bc
Reviewed-on: http://review.whamcloud.com/10924
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>