HYD-3377 Odd behaviour causes lnet state to bounce up and down
HYD-1848 implemented a situation where corosync state was
reflected in the lnet state.

This logic is no longer required since the new lnet code and
in particular the status of corosync is not related to the
status of lnet.

The user problem is that this can cause the reported lnet state to
go up -> down -> up. Corosync marks it down and lnet marks it up

Change-Id: Ibc0b114d47768acfc40a1ac4125026a603bd79a1
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/11105
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Aric Coady <aric.coady@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>