HYD-3403: Disable CLI test "Scenario: Force-remove a server"
We know from HYD-3211 that we have a low priority bug where we
can't force remove a server from the CLI. Due to
HYD-3400/TEI-2402, we didn't notice that there was a test failure
for this. Now that we're fixing these tests, we hit a test failure
for this bug. This ticket is for disabling this test until the bug
becomes high enough priority to fix.

Change-Id: I1938b5c18f7c8d53067f29563ff93510f08f4ca9
Reviewed-on: http://review.whamcloud.com/11250
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12786
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>