HYD-3404: Disable CLI test step in "Scenario: Reboot a server"
We know from HYD-3206 that we have a low priority bug where
Removing a server doesn't give a confirm prompt. Due to
HYD-3400/TEI-2402, we didn't notice that there was a test failure
for this. Now that we're fixing these tests, we hit a test failure
for this bug. This ticket is for disabling the test step checking
for the confirmation until the bug becomes high enough priority to
fix.

Change-Id: I8f6bbaab86f5f6daecb605219a558ce7bcf9fc5c
Reviewed-on: http://review.whamcloud.com/11251
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12787
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>