HYD-3398 IML needs a common directory for common utils and files.
Many routines could be shared between manager, simulator, agent etc
if we had a common place. This ticket will add that common place
and empty lib/util.py file and allow people to make use of it in
the future.

A small change to the build it required to add cause find to search
symbolic links.

The chroma_common modules are available from chroma_agent, chroma-director,
chroma_api, chroma_cli and chroma_core

Change-Id: Ic625f1a2246dd8404fd83731f401077bb4475dee
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/11230
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>