HYD-3428 Move shell.py into the common code directory.
shell.py which is a good candidate for common code needs to be
moved into the common code directory.

I'm doing this as a separate ticket as it is a large number of
very simple changes.

This does highlight some issues with the common-code namely that
references from within the common_code to other common_code need
to be relative or we get python namespace issues.

Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Change-Id: I6903b04156296845bf8d6a9d0ccb2a28bb41b801
Reviewed-on: http://review.whamcloud.com/11348
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>