HYD-3418 Prevent LNet config changes on copytools related to 'available' fs
LNet configuration is prevented in situations where the Copytool itself is related
(using a client mount) to a filesystem in the 'available' state.

Implementation leverages existing support in the UI code by enhancing the method in
host that computes if the host is associated with an available filesystem.

Also, took liberty to improve a method name.  Changed 'active' to 'available' in a backend method
to make that bit a bit more self documenting.

Change-Id: If5e8cd8aa3dff4294432cfb56e896c3623858a44
Reviewed-on: http://review.whamcloud.com/11358
Tested-by: Jenkins
Reviewed-by: Aric Coady <aric.coady@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>