HYD-3513 Create long-polling fit for 3.0
For reasons for performance and user experience we
must stop the 'gui' from polling the backend and use
a method o long polling to provide create more of a
push enviroment. Whilst the changes will be simple we
probably want to implement this is an incremental
roll-out across the end points.

The solution should be good for all cases but we can
limit the test to those used for 3.0

3.0 will mean

lnet configuration.
profiles selection for enhanced host addition.

 - Add a centalised blocking service for the database
   changes, this allows any iml-process to block on a
   change for a specific set of tables.
 - Extend job_scheduler_client to provide a client call
   to this centralised service.
 - Add tests to ensure the blocking works.

Change-Id: Idd59b1f3baa9cc651560cd9c368e712a903e3857
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17332
Tested-by: Jenkins
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>