HYD-3451 - Integrate UI errors into the realtime service
Sends UI error messages to the realtime service such that it can be
formatted with the correct file and line number and column. When the
formatted data is returned, the UI dialog will update with the new
formatted stack trace. This will only happen in a prod build, not in
development. It also checks to make sure that the stack trace includes
both line numbers and column numbers. Without both, it will skip sending
this info to the realtime service and instead display the standard error
message immediately.

Replaced HTTP_USER_AGENT with User-Agent.

chroma-externals: 11568/11

Change-Id: Ib58001fd34d1d65587192bf4d0fff584ae07054e
Reviewed-on: http://review.whamcloud.com/11569
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>