HYD-3462 Add support for running on EL 6.6
Create an /etc/cluster/cluster.conf file when collecting pcs config
output as it's now required. The file is just a placater for the pcs
command. pcs doesn't actually read it but wants it to be there with
some loose form of valid content.

Also adds 6.6 to the supported versions in the install script.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I41bca5c2aeee8b2e1b03542a4b9d152113e809e0
Reviewed-on: http://review.whamcloud.com/11635
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12522
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>