HYD-2592 Libraryize some common test code
As our test scripts have grown over time, we've ended
up with a redundant definition of a number of variables
that repeat themselves in every test script. This patch
moves the variables into a common file, defaults.sh, and
includes the file at the beginning of each script.

This definition in one shared location is a small step towards
removing the redundancy in our test scripts.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I33c70c25f791d458c2160ef5d2a6e1b7f4547775
Reviewed-on: http://review.whamcloud.com/11674
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>