HYD-3516 Allow a pragma to emulate a version tag
Because tags cannot be pushed with reviews and it's interesting
at times to see what the result of a tag is on the version of
iml-* packages produced and how they will upgrade, etc. emulate
what setting a tag will do using a commit message pragma.

Also, get the IEEL archive version from the IML version.  They are
the same thing now.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Icb27453dfc6dee77dd1c2154624747feee9c2f84
Reviewed-on: http://review.whamcloud.com/11760
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>