HYD-3432, HYD-3520 Properties are stored with the host for profile validation.
As part of the support for enhanced add host dialog additional
properties of the host are available to the manager so that
pre-install filtering might be applied. These properties could be
as examples memory size, whether zfs is installed, disk size, number
of cores etc etc.

Profiles can then be filtered on these properties so that ono compatible
ones are provided to the user as options.

Change-Id: I53f40db70a9d28f6bafece70d03998845bedfe61
Reviewed-on: http://review.whamcloud.com/11927
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>