HYD-3460 allow specification/selection of lanplus with IPMI interfaces
Minimalistically designed implementation to allow IPMI 1.5 or 2.0 (lanplus) for
an IML installed instance.  And extra record is included in the powercontrol types
fixture which is loaded into the DB at install time.  The user is asked which
type to use when ipmi is setup the first time.

All the power control types are loaded and cached on the page now.  This code
accesses the page case to get the IPMI power control types.

Change the integration testing setup to match on the new impl of
PowerControlType.display_name.  This new impl strips off (IPMI).

And now lanplus t/f to the agent so that the -P can be added to the fence_agent

Change-Id: Ic5b8221722651916aa42798bf51caf134e4c4c77
Reviewed-on: http://review.whamcloud.com/12032
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>