HYD-3777 Extend classes created for ZFS testing to allow testing of lvm and eventually mdraid
Some existing work contains additional improvements that were raised by
Kelsey in the zfs landing by Kelsey for ticket HYD-2615.

The extended the classes enabling testing of a number of backends to allow
basic testing for lvm and mdraid. These additional classes now mean that
during existing file system testing we test ldiskfs, zfs, lvm and
mdraid. The tests are simple because only one device is used for each
configuration - i.e. It is a raid array of 1.

The key piece of functionality here is that it will prevent a regress
of the ability to monitor simple lvm based systems.

NOTE: This patch currently doesn't enable mdraid because mdraid needs
to be validated in iml first.

Change-Id: I6b81653232aef91a12bc2d7c7c88aeb21ac7fffe
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12072
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>