HYD-3628 Existing filesystem tests fails to install lustre-osd-zfs
For some unexplained reason the lustre-osd-zfs started
failing to install unless the --enablerepo=lustre flag was
added. It's odd because the repo is enabled although
there are 2 'lustre' repos one of which is disabled
and so perhaps the outcome is indeterminate.

The enablerepo=lustre does no harm and perhaps further on
we should try to rename one of the lustre repos
something other than lustre.

Change-Id: Id935c4d449af335c7ec04f71a5763dd5e43aa58f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12089
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>