HYD-3627 Correct regression in populating chroma-externals
A regression was introduced in 564d7fd0 where "fetch-from-repo"
externals are no longer being downloaded by "make download".

This attempts to correct that, but in all honesty it has not been tested
against the externals that 564d7fd0 introduced and is only intended to
get the externals needed to close on feature freeze of IEEL 2.2.

Change-Id: Ie93eecbefbefa89a64e7dcab6a184ffe9b128dee
Reviewed-on: http://review.whamcloud.com/12094
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Tested-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>