HYD-3629 zfs 0.6.3 filesystem detection not working
Previous to this patch I had made of of the zfs debug
utility zdb to extra properties from the zpool and
zdataset. This turned out to be a bad solution because
'zfs get' works better and the output from zdb is not
defined and so liable to change. Hence the fact that
between zfs 0.6.2 and 0.6.3 IML stopped working.

This new patch works through the defined 'zfs get' and
contains a little bit of knock on refactoring due to
name changes.

Change-Id: I7e8614233cee53f75a38e1f9fdfcd99a5e79f396
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12156
Tested-by: Jenkins
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Tested-by: Joe Grund <joe.grund@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>