HYD-3643 Sets profiles is incomplete preventing installation of hosts
The code to set the profile at the intial default install was
incomplete on the first landing. This code moves someway towards
completion include the setting up of the host, installation and
removal of packages and reworking of the installation process so
that only the require parts of the system are configured.

A change to regularise the ExportedZfsDevice code, Mathieu made a
comment in the review and this is an opportunity for a simple tidy up.

A very simple change to cause the error when the properties are not
available for evaluation at host install to be informative rather than
a unknown variable error.

Change-Id: I0e6a330904a411b59aa2722b2ed95119e874720f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12187
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>