HYD-3707 Solution for the SSLv3 POODLE Security Issue
This fix disables SSLv3 on IML connections and includes
some test to ensure it is done so.

About POODLE.

On October 15th 2014, Google announced that they had
found a serious security flaw in an 18-year-old variant
of the SSL protocol known as SSL version 3, or SSLv3.

In their publication, Google describes how an attack
they've called 'Poodle' (or to give its formal designation,
'CVE-2014-3566') can be used to defeat the encryption in SSLv3.

Unlike the 'Heartbleed' security bug uncovered earlier
this year, this is not a vulnerability in any particular
implementation of the SSLv3 protocol, but rather a subtle
defect in the design of the SSLv3 protocol itself.

As a result, SSLv3 is now widely regarded as being
no-longer fit for the purpose of protecting sensitive
information as it travels over the Internet.

Change-Id: I9a36b54111a93217aaf83d5f779f97a5ed4a9d86
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12507
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>