HYD-3756 ZFS Unavailable pools cause zfs detection issues.
Update so that the zfs scan code does not scan zfs pools that
are listed by zfs as unavailable. The code changes are pretty
straight forward.

Along with this I added some much needed zfs testing, the zfs
scan code did not have any unit test and this is now resolved.

zfs requires the commands can return errors and this meant
improving the patch of try_run / run to return errors but also
return errors when a command is not found. This change meant
that manny other unittests in the agent had to be improved to
list all the commands that they use. This makes all this tests
more robust.

Change-Id: If5cac6d68b66de7049f0dc74ba4be7a4835cb2af
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12477
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>