HYD-3783 Intermittently incorrectly getting Incompatible for managed server profile
The problem in the ticket is caused by errors in the agent not being caught
by plugins but being allowed to cause the plugin to exit meaning a single error
in one part of a plugin can stop all data being returned.

This patch creates ExceptionSandBox's which localize unexpected exceptions and
their consequences. The method used is simple and allows the sand boxing to be
removed for debug / test environments. This means that in the debug situation
we will see the worst effect but in the production environment we should see a
more gradual reduction in capability.

This patch clears up the issues causing the HYD-3783 ticket, which is an error
in the audit plugin, it does not sandbox device capture etc. We will do that in
follow on patches once we have experience of this patch.

Change-Id: I7038895f2528802b892af7315d496251a2806826
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12755
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>