TEI-1479: Race reading from paramiko channel
We have a race with reading from stdout/sterr from a channel and garbage
collection perpetuated throughout our tests. _ssh_address is storing
the file-like objects created as part of the channel and returning those
attached to an object. But once we leave the function, the channel is free
to be garbage collected. When the channel is garbage collected, it closes
that channel. Per the Paramiko docs, attempts to read from the channel
after the channel stream has closed results in a zero-length string.
This is why we saw HYD-2912 where it seemed like Paramiko lied and
said a directory was empty when it wasn't - it was returning a zero
length string as the stream had closed. Instead, we need to read from the
channel before we leave the function and its available for garbage
collection, and should be closing the channel when we're done while we're
at it. This patch also removes the workaround added for HYD-2912 as it is
no longer needed as the actual problem with how we used paramiko has been
fixed.

Change-Id: I03498de1f7338f203fdbc69e8f7c70cc4cc4e4dd
Reviewed-on: http://review.whamcloud.com/12777
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>