HYD-3705 Configure LNet is not enforcing that the file system be stopped before editing
The check for member_of_available_filesystem was to see if the
filesystem was available, however a partially active filesystem
can be available or unavailable. We only what the user to edit
lnet configuration if the filesystem is actually stopped. So
make the test for an active filesystem not availabe filesystem
(renaming variables as appropriate) and then increase the filter
to include state in (available, unavailable).

Change-Id: I7f66f3edd745e4c62ec1e9720cd4f97d430b9e55
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12770
Tested-by: Jenkins
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>