HYD-3717: Make integration tests use new server addition
We need to convert our integration tests to all use the new server addition
code path added with IEEL 2.2. The fact our main system testing is not using
the same method to add servers as nearly all or all of our customers will be
using is a major gap in coverage. This can be accomplished by updating the
add_hosts function in ChromaIntegrationTestCase. The upgrade test needs to
still be able to use the old method of adding hosts when interacting with
older versions of IEEL that do not have the new method available.

Change-Id: I8f3e7e761e4fcbb170b88c80e5a8541a063cce00
Reviewed-on: http://review.whamcloud.com/12778
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>