HYD-3860 Add git hooks to warn if the chroma externals ref. has been modified.
It is quite easy to accidentally modify the chroma-externals reference
by accident.

Create a prepare-commit-msg hook to add a warning at the top of the
commit message so that the committer is aware such a change has been
made.

To make the script active copy the prepare-commit-msg from scripts to
.git/hooks

Change-Id: Ic4863a9c1b721ab2cb64c21827b1a4a31272855d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12796
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>