HYD-3880 Reduce incidence of issues caused by corosync using a bound port
The port chooser for the corosync configuration can choose
existing ports. This patch does not resolve that issue but
does reduce the range used for searching from 0-65535
to 32767-65535. This is an interim change to reduce the
likelihood of collision.

It is far to say the just looking about 1024 would have reduced
the likelihood of a collision greatly but from some experience
I think ports are generally at lower rather than higher numbers.

Change-Id: I31a1b9668b391aed8db09b38f2eed0dc0f560eca
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12847
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>