HYD-3861 Make the create_lustre_filesystem code always use index
--index is optional for all creating mgt/mgs accept we seem to see
cases where Lustre requires it. This may be a subtlety about Lustre
or about our setup. Using --index is correct and so the test framework
will be more reliably if it does.

This patch adds --index to the creation of all filesystem elements, and
makes it a requirement to use it in the applicable library routines.

Change-Id: I34855232c9a039f775cdc469c78b167f4012bc81
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12863
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Michael MacDonald <michael.macdonald@intel.com>