HYD-3660 Stats sample tables are growing large with job stats
This change allows for 2 methods of deleting expired statistics. The original
method which deletes expired per id as they are added. This has the problem
that when a statistic stops being added the old points stop being deleted
which for jobstats in particular is not good.

This new method just deletes all of values once they have expired regardless
of their id. It does this occasionally which should reduce the sql being executed
as well.

The change also changes the retention times to something more useful.

Change-Id: Ib71f3b22d1968663c6c2f43b15110dbc5d5df813
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/12979
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Michael MacDonald <michael.macdonald@intel.com>