HYD-3659: Devices with partitions cause exception on cull
When a removable device with partitions on it is removed, the
manager will attempt to cull the created Volume and other
associated entities from the ResourceManager. Due to the odd
nature of a Partition (it is both a global resource and a
dependent of the base LinuxDevice), the cull routine first
deletes the partition as a dependent and then attempts to
delete it again as a global resource.

This patch adds some logic to run the queries for lost resources
in a loop. If a resource is deleted as a dependency of prior resource,
it won't show up in the next iteration's query.

Change-Id: Ia9b18f9d0dd2c690782153bcf2e05c44801ef1e3
Reviewed-on: http://review.whamcloud.com/13036
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>