HYD-3924 Mock agent not reporting any properties causing selenium installs to fail.
The host properties are complete fakes for the mock and due to the locking in the
job schedular even those fakes were not being saved. This change as a different
approach sets up a django hook so that it can set the properties and the HA pairs
during the added at that point.

This removes the workaround for this bug that was added as part of HYD-3923.

During this work I found a couple of duplicate functions in the mock helper and these
were removed.

Change-Id: Ifee8ffc1d0d88e1ed002123561047c5f306184d4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13085
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>