HYD-3935 More efficient client count computation
The solution for HYD-2916 uses cat in a loop. What this means is
that every 10 seconds or so, the agent has to fork/exec N cat
processes, where N is the number of exports under the MDT's /proc
directory.

This patch uses the built-in methods supplied by the audit classes
to do the computation more efficiently in the main agent process.

Change-Id: Ic8e4897ed2b97d670e3a0cfc56ff8ae6d76f5d6c
Reviewed-on: http://review.whamcloud.com/13091
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>