TEI-3015: Add debug info to IML test reporting
Currently if there is an issue with the test reporting jobs
there is little to no information to see what went wrong.
Firstly, lets print what we sent to Gerrit, so we know if
we even sent a message to it or not. Secondly, lets fix the
logic around printing the output from the aggregation script
so that it actually prints the output on failure and not
just success. Thirdly, lets archive a test_info.txt from
each test job so that Jenkins links the jobs even if
no test results were generated by the upstream test job.

Change-Id: I5a73a6eb15e43c224930ba7897762e1694d6f1c2
Reviewed-on: http://review.whamcloud.com/13099
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>