HYD-3186 Improving scaling by making Lustre audit only send deltas
This ticket makes a change to the audit code to only send the data
for each component if something in the component has changed. It isn't
supersmart so if any mount changes (for example) all the mount data is
resent. But most of this data changes very slowly.

The receiver is modified to handle the fact that data is only sent on saved.

It doesn't attempt to reduce the chatter of the stats arriving by aggregating
or not reporting stats which have no activity since the last report.

As a catch all data is always sent every 10 minutes, changed or not, so that
if we have a flaw in the algorithm and something gets missed updates eventually
occur. But I would suggest the the code is more likely to send unnecessary changes
than miss actualy changes.

Testing was added to check for the change discovery, and because I was in that
space I broke much of the lustre audit testing into individual files rather than
keep the big multi target test file.

Finally as a knock on in the event of crm_mon errors we report increased logging
giving us improved debug-ability.

Change-Id: I999d4d8695b7c614e26a8b77bc039ee50094d156
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13120
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Michael MacDonald <michael.macdonald@intel.com>