HYD-3943 Test failure: unit, chroma_core, models, test_samples, test_stats, list contents too low
When HYD-3660 was implemented I (Chris) tried to make the test more sophisticated
by making the number of points exactly fill the highest resolution entries. This meant
that the test below for sample counts would test a corner case of some points being
expired and some points being preset (I presumed that was what was intended). However
the behavour at that edgeis difficult to understand and so needs to be resolved as part
of HYD-3943. To lower the status of HYD-3943 from blocker I am setting the number of
points back to the value previous to HYD-3660 which was 100. I will then close HYD-3943
but haved opened HYD-3960 to track the fix of the possible underlying issue.

Change-Id: I7f76ac34000fe37ced8c097950db22a5666eef7d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13234
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>