HYD-4005: Fix upgrade testing on RHEL.
RHEL testing has regressed while it was disabled due to
HYD-3859, preventing re-enabling. We need to get it back
to a passing state so we can turn it on and prevent
further regression. This is achieved by starting on a
RHEL 6.6 compatible starting version since there is
no real concept of RHEL 6.5 anymore since RHN updated
its repos to 6.6 packages at GA.

Change-Id: Ib2dc0c960defeaf688bee39a4daf443e202b7282
Reviewed-on: http://review.whamcloud.com/13355
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>