HYD-3988 Workaround segfault in python in EL 6.6
This rather strange patch is a result of HYD-3988. The
segfault SEEMS to be isolated to the case where 'service
ntpdate restart' is run by shell._run . If as is the case
here 'service ntpdate restart' runs whilst shell._run is
executing then the segfault does not occur. This is
evidenced by the fact that the patch runs 'service
ntpdate restart' delayed and in a completely independent
shell and then separately runs another command
to detect that it was run. This doesn't cause a segfault.
At some point once the root cause is resolved the patch
should be reverted back to some reasonable code

Change-Id: I6926c7d708b0c8d980b2f564649335f3f26e3179
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13367
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>