HYD-4005: Workaround paramiko bug #448
In the current version of paramiko we are using, there
is a bug that causes channel.recv_exit_status() to hang
indefinitely when the stdout for the command exceeds a
certain size. This is causing an issue in our upgrade
testing as the yum -y update from RHEL 6.5 up to RHEL 6.6
is generating a lot of output, causing the tests to hang.
While this is fixed in later versions of paramiko, it is
not an appropriate time to upgrade in our release cycle.
We can, however, workaround the issue by using this patch,
and it will still be correct even with the bug fix in later
versions of paramiko.

Change-Id: I673374e2e0fdaa85b366b035da31c02d75fe87e2
Reviewed-on: http://review.whamcloud.com/13394
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>