HYD-4022 Get client kernel from the kernels repo
When installing client and worker nodes in CI the required Kernel 
for the Lustre client is not always present in the standard
distribution files.

A working example is when trying to a install Lustre client that
requires an el6.6 Kernel on a el6.5 installation.

By adding the reference to a cobbler repository that contains
additional Kernels to the standard distribution Kernels we enable
the Lustre client to install on these nodes.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ib3252b5853d687b8d8a75283867bfcc47de4e3fc
Reviewed-on: http://review.whamcloud.com/13385
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>