HYD-4024 fstab can contain UUID which blkid cannot read
UUID for devices in fstab may not be valid or may cause blkid
to throw errors when used as an argument. The reasons may be
numerous but iml should be more resilient to such occurances.

This patch makes IML more resilient by catching the error state
when a UUID in the fstab file cannot be resolved into a path.

UUID's that can't be resolved to a patch cannot be used by IML
and so excluding them from results is the best behaviour. A
warning is logged so the lack of translation is recorded.

Change-Id: I8a9d5eef5446f602e0397178673387efa4b9c186
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23212
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>