HYD-3684 updates available alert raised on freshly added servers
The lustre bundle comes with kernel-headers as well as lustre and
other artifacts. However the when lustre is install it's dependencies
to not include kernel-headers. They include the kernel and so that is
installed but nothing requires the kernel-headers.

So after completion everything is up to date except the kernel-headers
which were never upgraded.  If Kernel-Headers where not installed on
the original machine then an update would not be indicated as no
packages would be out of date.

The solution is to run yum check-update with only the IML repos enabled
and then run a yum upgrade with the distro repos enabled and explicity
naming the packages found from the yum check-update

This patch contains a small fixup in TestMdtMetrics.test_get_client_count
which was created by a previous patch http://review.whamcloud.com/#/c/13449/
upon which this patch depends.

Change-Id: I6f0418676c10af517f1ae3c9cfd61b4e6f0c81cb
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13450
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>