HYD-4052 test_lnet_operational_after_failover: AssertionError: 0 not greater than or equal to 4 on EL 6.6
After installed a number of hosts it is a requirement to
wait a little before the common resources between them common
meaning HA capable, to be resolved by IML.

This patch means that we wait for this to happen rather than
require it to have been available immediately.

Review http://review.whamcloud.com/#/c/13313/ is a general solution
to this but I am having problems getting it to generally work
so because this is a blocker I would suggest landing it
while the general solution is resolved.

Change-Id: I5ba4c9ef504127af173535365347de32f48974a5
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13459
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>