HYD-4054 Can't add storage server using command line
The CLI cannot access the database and so any code must only use the rest interface.

Change the access of the job and step information to use the rest interface rather
than access the database directly.

This change also highlighted a bug in one of the mocks which was creating a
generic Job object rather than the required TestHostConnectionJob object

Another change required was because the list of checks in test host had
not been updated causing a key error on the return.

Finally because test-host requires the command to complete we have to manually
wait for the command to complete even if no wait is set. The CommandMonitor
class was extended to allow for this.

Change-Id: I3340030a0608e1481951b049f88e7710133974d6
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13466
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>