HYD-4096 Redeploy Agent does not work
When a node fails to deploy, the backend code incorrectly expects the
host to be passed in a field called 'host',
when the protocol defines the field as 'address'.
This causes an error to be raised and returns to the user as 
a stack trace.

This patch modifies the backend to expect the field as 'address' instead
of 'host'.
It also changes the front end to always send the default profile, whether it's a put or post to /host.


  - Check address in data instead of host.
  - Make sure front end always sends default profile.

Change-Id: I4cd38b60a3777f17e25f271d149049ca3d21e829
Reviewed-on: http://review.whamcloud.com/13484
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>